refactor(pages/home): split webinar section to Webinar.tsx
